Final Upload for Sem1 WT project .
index.html is the main html page where you can search for images. It is connected with index.js for api and style.css for css. In the same way, Contact.html has its own css and js and about.html tell us about the code how it works and its styling is done in css